+++ .gitignore
@@ -1,2 +1,2 @@
-private.env
+.env
 node_modules
