+++ README.md
@@ -1,20 +1,7 @@
 # SoulCaptain
-The first step of making your dream a reality is to define and share it.
-
-We have a dream of building an application for dreaming your dream. 
-
-## [run app](https://summer-name.glitch.me/)
-
-## [slack](https://codemoon.slack.com/messages/C08J04G7Q/details/)
-
-## [wiki](https://gitlab.com/maciejjankowski/soulcaptain/wikis/home)
-
----
 
 # todo
 
-## [BOARD](https://gitlab.com/maciejjankowski/soulcaptain/boards/)
-  
 ### Maciej 
 * [ ] odhaczanka
   * [ ] ustalenie z Wiktorem atrybutów soulencji odhaczanki - nie może to być "soulType" bowiem można dodać do odhaczanki np. odmówić modlitwę, przeczytać afirmacje, więc albo zrobimy podtyp suolencji albo np. bool "doable" "yes|no"
@@ -27,18 +14,11 @@ We have a dream of building an application for dreaming your dream.
 * [ ] walidacja danych https://github.com/epoberezkin/ajv
 * [ ] opcjonalnie facebook / gmail login - uprawnienia do api w google console 
 * [ ] opcjonalnie facebook / gmail login - https://github.com/passport/express-4.x-facebook-example/blob/master/server.js
+* [ ] soulencje -> karty manifestu: http://jsbin.com/qekagiq/25
+* [x] Logowanie i apmiętanie usera w sesji
   
 ### [Wiktor](https://gitlab.com/maciejjankowski/soulcaptain/boards?scope=all&utf8=%E2%9C%93&state=opened&assignee_username=roktiw)
-* [ ] widoki logowania (login.html) i zakładania konta (signup.html) przepisać na [floating-labels](https://getbootstrap.com/docs/4.0/examples/floating-labels/)
-* [ ] [czyta](https://keep.google.com/share?tid=true#NOTE/1516399000064.111084.3791535080)
-* [ ] [formularz dodawania soulencji](https://jsbin.com/sepupuw/edit?html,js,output)
-* [ ] widok karty w Bootstrap
-* [ ] zaznaczanie klasu active dla elemenu menu po urlu
-* [ ] wydzielenie MENU do partiala
-* [ ] widok decku tinder style
-** https://codepen.io/ionic/pen/nxEdH?page=1&
-** https://admiralpotato.github.io/vue-swipe-cards/ <-- na vue!
-* [x] widok decku w Bootstrap
+
 
 ---
