+++ api/habits.js
@@ -4,23 +4,27 @@ var timeSpans = ['hourly', 'daily', 'weekly', 'monthly', 'quarterly', 'annually'
 var habitType = ['growth', 'excite', 'sustain', 'maintenance', 'challenge']
 
 const Habit = mongoose.models['Habit'];
-  
+  /*
+  expect(req.body.payload).to.beArray()
+  expect(req.body.payload[0]).to.have(idProperty)
+  */
   app.post('/habits', (req, res) => {
       if (req.body && req.body.payload && req.body.payload.length){
         let count = req.body.payload.length;
           req.body.payload.forEach((habit) => {
             console.log('habit', JSON.stringify(habit,null,2))
               Habit.findOne({"id" : habit.habitId}).then((foundHabit) => {
+                // /* expect(foundHabit).not.to.be.empty() */
                 if (foundHabit){
                   foundHabit.lastCompleted.push(Date.now());
                   foundHabit.save().then(()=>{
                     count--;
-                    if (!count)  
+                    if (!count)
                       res.send("OK");
                   }).catch((err) => {
                     console.log('error saving', err);
                     count--;
-                    if (!count)  
+                    if (!count)
                       res.send('ERR');
                   });
                 } else {
@@ -29,13 +33,17 @@ const Habit = mongoose.models['Habit'];
                   let newHabit = new Habit(habit);
                   newHabit.lastCompleted = [Date.now()];
                   newHabit.save().then(()=>{console.log('saved')});
-                  if (!count)  
+                  if (!count)
                     res.send("OK");
                 }
-                  
               });
-            
           })
       }
-  });
-}
+  }); // POST /habits
+
+	app.get('/habits', (req, res) => {
+    console.log(req.session.user);
+		res.send(JSON.stringify({session : req.session, user : req.user }));
+	})
+
+};
