+++ package.json
@@ -1,44 +1,46 @@
 {
-	"//1": "describes your app and its dependencies",
-	"//2": "https://docs.npmjs.com/files/package.json",
-	"//3": "updating this file will download and update your packages",
-	"name": "my-hyperdev-app",
-	"version": "0.0.2",
-	"description": "What am I about?",
-	"main": "server.js",
-	"scripts": {
-		"start": "node server.js"
-	},
-	"dependencies": {
-		"express": "^4.16.2",
-		"mongoose": "^4.13.9",
-		"passport": "^0.4.0",
-		"passport-local": "^1.0.0",
-		"passport-google-oauth": "^1.0.0",
-		"passport-facebook": "^2.1.1",
-		"connect-ensure-login": "^0.1.1",
-		"express-session": "^1.15.6",
-		"body-parser": "^1.18.2",
-		"cookie-parser": "^1.4.3",
-		"connect-mongo": "^2.0.1",
-		"request": "^2.83.0",
-		"nodemailer": "^4.4.1",
-		"mailgun-js": "^0.14.1",
-		"mustache-express": "^1.2.5",
-		"bcrypt": "^1.0.3",
-		"smsapi": "^1.5.6"
-	},
-	"engines": {
-		"node": "6.10"
-	},
-	"repository": {
-		"type": "git",
-		"url": "https://hyperdev.com/#!/project/welcome-project"
-	},
-	"keywords": [
-		"node",
-		"hyperdev",
-		"express"
-	],
-	"license": "MIT"
+  "//1": "describes your app and its dependencies",
+  "//2": "https://docs.npmjs.com/files/package.json",
+  "//3": "updating this file will download and update your packages",
+  "name": "my-hyperdev-app",
+  "version": "0.0.2",
+  "description": "What am I about?",
+  "main": "server.js",
+  "scripts": {
+    "start": "node server.js",
+    "watch": "nodemon server.js"
+  },
+  "dependencies": {
+    "bcrypt": "^1.0.3",
+    "body-parser": "^1.18.2",
+    "connect-ensure-login": "^0.1.1",
+    "connect-mongo": "^2.0.1",
+    "cookie-parser": "^1.4.3",
+    "dotenv": "^5.0.0",
+    "express": "^4.16.2",
+    "express-session": "^1.15.6",
+    "mailgun-js": "^0.14.1",
+    "mongoose": "^4.13.9",
+    "mustache-express": "^1.2.5",
+    "nodemailer": "^4.4.1",
+    "passport": "^0.4.0",
+    "passport-facebook": "^2.1.1",
+    "passport-google-oauth": "^1.0.0",
+    "passport-local": "^1.0.0",
+    "request": "^2.83.0",
+    "smsapi": "^1.5.6"
+  },
+  "engines": {
+    "node": "6.10"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://hyperdev.com/#!/project/welcome-project"
+  },
+  "keywords": [
+    "node",
+    "hyperdev",
+    "express"
+  ],
+  "license": "MIT"
 }
