+++ server.js
@@ -1,11 +1,9 @@
 const express = require('express');
 const app = express();
 
+require('dotenv').load();
 const mongoose = require('mongoose');
 const passport = require('passport');
-const FacebookStrategy = require('passport-facebook').Strategy;
-const GoogleStrategy = require('passport-google-oauth20').Strategy;
-const LocalStrategy = require('passport-local').Strategy;
 const mustacheExpress = require('mustache-express');
 const session = require('express-session');
 const MongoStore = require('connect-mongo')(session);
@@ -13,13 +11,17 @@ const mail = require('./email.js');
 const bcrypt = require('bcrypt');
 const sms = require('./sms.js');
 
-
-mongoose.connect(`mongodb://${process.env.DBUSER}:${process.env.DBPASS}@${process.env.DBHOST}/${process.env.DBNAME}`, { useMongoClient: true });
+const mongoConnString = `mongodb://${process.env.DBUSER}:${process.env.DBPASS}@${process.env.DBHOST}/${process.env.DBNAME}`;
+mongoose.connect(mongoConnString, { useMongoClient: true });
 mongoose.Promise = global.Promise;
 
 
-app.engine('mustache', mustacheExpress());
-app.set('view engine', 'mustache');
+app.engine('html', mustacheExpress({tags : [ '<%', '%>' ]}));
+app.set('view engine', 'html');
+app.set('view options', {tags : [ '<%', '%>' ]});
+app.locals.delimiters = '<% %>';
+app.locals.tags = '<% %>';
+
 app.use(express.static('static'));
 app.use(require('cookie-parser')());
 app.use(require('body-parser').urlencoded({ extended: true }));
@@ -28,218 +30,28 @@ app.use(require('body-parser').json());
 app.use(session({
   secret: process.env.SESSIONSECRET,
   store: new MongoStore({ mongooseConnection: mongoose.connection }),
-  resave: true, saveUninitialized: true 
+  resave: true, saveUninitialized: true
 }));
 
 app.use(passport.initialize());
 app.use(passport.session());
 
-// require('./api/habits.js')(app, mongoose);
-
 // using passport: https://stackoverflow.com/questions/45381931/basics-of-passport-session-expressjs-why-do-we-need-to-serialize-and-deseriali
+const schema = require('./api/schema.js')(mongoose);
 
-passport.use(new LocalStrategy({ usernameField: 'email' }, (email, password, done) => {
-  User.findOne({ email: email.toLowerCase() }, (err, user) => {
-    if (err) { return done(err); }
-    if (!user) {
-      return done(null, false, { msg: `Email ${email} not found.` });
-    }
-    console.log('passport login', password, user.password);
-    testPassword(password, user.password, (err, isMatch) => {
-      if (err) { return done(err); }
-      if (isMatch) {
-        console.log('password matches!')
-        return done(null, user);
-      }
-      return done(null, false, { msg: 'Invalid email or password.' });
-    });
-  });
-}));
+require('./api/passport.js')(mongoose, passport);
+require('./api/login.js')(app, passport);
+require('./api/deck.js')(app, mongoose);
 
-passport.serializeUser((user, done) => {
-  done(null, user.id);
-});
-
-passport.deserializeUser((id, done) => {
-  console.log('deserialize', id);
-  User.findOne({ email:id }, (err, user) => {
-    done(err, user);
-  });
-});
-
-app.post('/login', (req, res, next) => {
-  console.log('login attempt');
-  passport.authenticate('local', (err, user, info) => {
-    if (err) { return next(err); }
-    if (!user) {
-      console.log('no user', user, info);
-      return res.redirect('/login.html');
-      
-    }
-    
-    req.login(user, function(error) {
-        if (error) return next(error);
-        console.log("Success! Redirecting to /");
-      
-        return res.redirect('/');
-    });
-    //res.redirect('/');
-    console.log("info", info);
-
-  })(req, res, next);
-});
-
-const Card = mongoose.model('Card', // http://mongoosejs.com/docs/guide.html
-{  
-   "coulCardId" : String,
-   "soulCardTitle" : String,
-   "soulCardSoulencje" : [  
-      {  
-         "soulIdParent" : String,
-         "soulType" : String,
-         "language" : String,
-         "text" : String,
-         "source":{  
-            "author" : String,
-            "created" : Date,
-            "source" : String,
-            "sourceLink" : String
-         }
-      }
-   ],
-  owner : String
-});
-
-const Habit = mongoose.model('Habit', {
-    "frequency": String,
-    "interval" : Number,
-    "name": String,
-    "times": Number,
-    "type": String,
-    "lastCompleted": [Date],
-    "habitId" : String
- });
-
-const User = mongoose.model('User', {  
-  "loginId" : String,
-  "loginType":String,
-  "email" : String,
-  "password":String,
-  "personalInfo" : [  
-     {  
-       "firstName" : String,
-       "lastName" : String,
-     }
-   ],
-  "habits": [{ type: mongoose.Schema.Types.ObjectId, ref: 'Habit' }]
-});
-
-  
-
-var listener = app.listen(process.env.PORT, function () {
+var listener = app.listen(process.env.PORT || 9000, function () {
   console.log('Your app is listening on port ' + listener.address().port);
 });
-
-
-
-
-
-app.get('/deck',(req, res) => {
-  Card.findOne().then((cards) => {
-    res.send(cards);
-  });
-});
-
-app.get('/sms',(req, res) => {
-  // sms("Eco", "791813568", "cześć soulcaptain", ()=>console.log('poszło'), ()=>console.log('nie poszło'));
-  // schedule failure retry to background processsing
-  res.send('ok')
-});
-
-app.get('/testx',(req, res) => {
-  let data = {
-      testField : "testValue"
-  }
-  res.send('OK');
-  res.render("password", data, function(err, body){
-    if (err) return console.error(err);
-    mail('', 'subject', body, 'this is just a text');
-  });
-});
-
-
-app.post('/card2', (req, res) => {
-  var cardData = {};
-  var card = new Card(cardData)
-  
-  card.save(function (err) {
-    if (err) {
-      console.log(err);
-      res.send(400, {status:'error',error: 'problem saving', details : err})
-    } else {
-      res.send({status:'ok'});
-    }
-  }); // card save  
-}) 
-
-app.get('/habits', (req, res) => {
-  res.render('habits')
-})
-
-app.post('/deck/:id/soulencja', (req, res)=>{
-
-})
-
-
-app.post('/user', (req, res)=>{
-  let fields = req.body.payload;
-  let userForm = {};
-  let validFields = "email password firstName".split(" ");
-  Object.keys(fields).forEach((index) => {
-    // console.log(field, field.name, field.value)
-    if (validFields.indexOf(fields[index].name) > -1){
-        userForm[fields[index].name] = fields[index].value;
-    }
-  });
-  
-  User.findOne({loginId : userForm.email}).then((user)=>{
-    if (user){
-      console.log(user)
-      testPassword(user.password, (err)=>{
-        if (err) return console.log('no matchy');
-        console.log('login ok', user);
-        res.send('OK')
-      });
-    } else {
-      bcrypt.hash(userForm.password, 10, function(err, pwhash) {
-        let newUser = new User({
-          loginId : userForm.email,
-          loginType : "email",
-          email : userForm.email,
-          password : pwhash
-        }).save();
-      });
-    };
-  });
-  
-  res.send("OK")
-}) // post user
-
-app.get('/profile',
-  require('connect-ensure-login').ensureLoggedIn(),
-  // isAuthenticated,
-  function(req, res){
-    console.log('/profile');
-    res.send({ user: req.user });
-  });
-
-app.get('/login', passport.authenticate());
 
 /**
  * Login Required middleware.
  */
 function isAuthenticated(req, res, next){
-  
+
   console.log('testing for login');
   if (req.isAuthenticated()) {
     console.log('login ok');
@@ -247,16 +59,14 @@ function isAuthenticated(req, res, next){
   } else{
 
     console.log('missing login');
-    
+
     if (req.headers['content-type'] === "application/json; charset=UTF-8"){
       res.status(403).send("please log in");
     } else {
       res.redirect('/login.html');
     }
   }
-};
-
-
+}
 
 app.isAuthenticated = isAuthenticated;
 /**
@@ -275,16 +85,5 @@ exports.isAuthorized = (req, res, next) => {
 require('./api/habits.js')(app, mongoose, isAuthenticated);
 
 
-function testPassword(pass, hash, cb){
-  bcrypt.compare(pass, hash, function(err, res) {
-    if(res) {
-      cb(null, 1);
-      console.log('match');
-    } else {
-      cb(1);
-      console.error('no match');
-     // Passwords don't match
-    } 
-  });
-}
+
