+++ static/deck.html
@@ -83,6 +83,12 @@
                                             </p>
                                             <hr>
                                             <div v-for="encja in card.soulCardSoulencje">
+                                              <div v-if="encja.soulType==='habit'">
+                                                habit ftw!
+                                                <a class="btn btn-primary">
+                                                  <i class="fa fa-check"></i>
+                                                </a>
+                                              </div>
                                                 <p class="card-text">{{encja.text}}
                                                 </p>
                                                 <div class="d-flex justify-content-between align-items-center">
