+++ static/habits.html
@@ -11,7 +11,9 @@
      <!-- JS for navbar -->
 
 <link rel="icon" href="../../../../favicon.ico">
-  
+
+</head>
+<body>  
       <!-- navbar start-->
 
     <nav class="navbar fixed-top navbar-light bg-light navbar-expand-lg navbar-light bg-light">
@@ -66,9 +68,7 @@
       cursor:pointer
     }
   </style>
-</head>
 
-<body>
   <div id="habits" class="container">
       <span v-on:click="editableHabits()" class="badge badge-primary btn-sm pointer">
       <i class="fa fa-pencil-square-o" aria-hidden="true"></i> Edit</span>
@@ -124,7 +124,6 @@
     <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.12.9/umd/popper.min.js" integrity="sha384-ApNbgh9B+Y1QKtv3Rn7W3mgPxhU9K/ScQsAP7hUibX39j7fakFPskvXusvfa0b4Q" crossorigin="anonymous"></script>
     <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0-beta.3/js/bootstrap.min.js" integrity="sha384-a5N7Y/aK3qNeh15eJKGWxsqtnX/wWdSZSKp+81YjTmS15nvnvxKHuzaWwXHDli+4" crossorigin="anonymous"></script>
     <script src="/habits.js"></script>
-
 
 </body>  
 </html>
